requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
qdrant-client==1.12.1
pydantic==2.10.6
python-dotenv==1.0.1
pandas==2.2.3
pyarrow>=14.0.0
numpy==2.2.2
sentence-transformers==3.3.1
transformers==4.47.1
torch==2.5.1
requests==2.32.3
pytest==8.3.4
